Fix Pagination in example raw code
Page 0 is the same as page 1. And the way that the page is incremented just before the check to see if you are at the final page causes the loop to exit early. By incrementing at the beginning of the loop you get all pages of data and don't miss the last page.