update environment variables to be prefixed with `CLOUDFLARE_`
Updates all references and usage of environment variables from `CF_`
prefix to `CLOUDFLARE_` should they be present. This follows suit of the
other SDKs and tooling in the Cloudflare ecosystem allowing reuse of
credentials between them all.

A full prefix was chosen to avoid conflicts and potential leaking of
details between other tools that use the `CF_` prefix.

Eventually, we will aim to deprecate the `CF_` lookups entirely but for
now, just prefer them if they are present.